+++ HISTORY.rst
@@ -127,6 +127,18 @@
 - Show cache partitions and partition file totals. This helps determine how
   effective is a cache maximum size value by showing how many objects and
   files the cache size limit is able to yield.
+- Statistics updates:
+
+  - Improve statistics navigation.
+  - Add doughnut chart type statistic.
+  - Add pie chart type statistic.
+  - Add column displaying the chart type per statistic.
+  - Update chartjs from version 2.8.0 to 3.9.1.
+  - Unify chart templates.
+
+- Add three document pie chart statistics: document count per document type,
+  document file count per document type, document file page count per
+  document type.
 
 4.3.1 (2022-08-21)
 ==================